Add (partial) Gingerbread support
Done this a few months ago, but it was reported that it only works
after tapping 4 times.

I’m guessing it is something to do with “onTouchEvent”. However,
swapping it to “dispatchTouchEvent” made the module have no effect.

I thought I’ll just commit this for anyone who wants to back port it
fully